Move auth scheme and its dependencies to a module.
