+++ test/auth_scheme.test.ts
@@ -1,8 +1,12 @@
 // Copyright (c) 2023 Cloudflare, Inc.
 // Licensed under the Apache-2.0 license found in the LICENSE file or at https://opensource.org/licenses/Apache-2.0
 
-import { TOKEN_TYPES, WWWAuthenticateHeader } from '../src/index.js';
-import { TokenChallenge, TokenPayload } from '../src/tokenBase.js';
+import {
+    AuthenticatorInput,
+    TOKEN_TYPES,
+    TokenChallenge,
+    WWWAuthenticateHeader,
+} from '../src/index.js';
 import { hexToString, hexToUint8, testSerialize, uint8ToHex } from './util.js';
 
 // Test vectors generated by pat-go (https://github.com/cloudflare/pat-go/)
@@ -32,16 +36,16 @@ test.each(tokenVectors)('AuthScheme-TokenVector-%#', async (v: TokenVectors) => 
     testSerialize(TokenChallenge, challenge);
 
     const context = new Uint8Array(await crypto.subtle.digest('SHA-256', challengeSerialized));
-    const payload = new TokenPayload(
+    const authInput = new AuthenticatorInput(
         TOKEN_TYPES.BLIND_RSA,
         TOKEN_TYPES.BLIND_RSA.value,
         nonce,
         context,
         keyId,
     );
-    const payloadEnc = payload.serialize();
+    const authInputEnc = authInput.serialize();
 
-    expect(uint8ToHex(payloadEnc)).toBe(v.token_authenticator_input);
+    expect(uint8ToHex(authInputEnc)).toBe(v.token_authenticator_input);
 });
 
 type HeaderVectors = (typeof headerVectors)[number];
